whisperdire/requirements.txt
openai-whisper==20231117
pyannote-audio==3.1.1
torch>=2.0.0
torchaudio>=2.0.0
pydub==0.25.1
numpy>=1.21.0
pandas>=1.3.0 